+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * refactor: removed support for https server
 * refactor: renamed Fritter.httpServer to server
 * refactor: removed local utility functions
+* refactor: made CorsMiddleware return its options
 
 ## 2.1.0